chore: separate config, lib and data
