+++ app/controllers/api/v1/quote_checks_validation_error_details_controller.rb
@@ -44,7 +44,7 @@ module Api
       def quote_check
         @quote_check ||= QuoteCheck
                          .select(*(QuoteCheck.column_names - %w[
-                           text anonymised_text
+                           text anonymized_text
                            file_text file_markdown
                          ]))
                          .accessible_for_source(api_user)
