+++ app/models/concerns/quote_input_metadata.rb
@@ -1,47 +1,11 @@
 # frozen_string_literal: true
 
 # Add profile and metadata from user inputs
-module QuoteInputMetadata # rubocop:disable Metrics/ModuleLength
+module QuoteInputMetadata
   extend ActiveSupport::Concern
 
   PROFILES = %w[artisan particulier conseiller].freeze # Also called Persona
   DEPRECATED_PROFILES = %w[mandataire].freeze
-
-  # Complete list of Geste types from QA prompts
-  GESTE_TYPES_GROUPS = {
-    "Chauffage" => %w[
-      chaudiere_biomasse
-      systeme_solaire_combine
-      poele_insert
-      pac_air_air
-      pac_air_eau
-      pac_hybride
-      pac_eau_eau
-    ],
-    "Eau chaude sanitaire" => %w[
-      chauffe_eau_solaire_individuel
-      chauffe_eau_thermo
-    ],
-    "Isolation" => %w[
-      isolation_thermique_par_exterieur_ITE
-      isolation_thermique_par_interieur_ITI
-      isolation_comble_perdu
-      isolation_rampants_toiture
-      isolation_toiture_terrasse
-      isolation_plancher_bas
-    ],
-    "Menuiserie" => %w[
-      menuiserie_fenetre
-      menuiserie_volet_isolant
-      menuiserie_fenetre_toit
-      menuiserie_porte
-    ],
-    "Ventilation" => %w[
-      vmc_double_flux
-      vmc_simple_flux
-    ]
-  }.freeze
-  GESTE_TYPES = GESTE_TYPES_GROUPS.values.flatten.freeze
 
   RENOVATION_TYPES = %w[geste ampleur].freeze
 
@@ -69,7 +33,7 @@ module QuoteInputMetadata # rubocop:disable Metrics/ModuleLength
 
     def metadata_values
       I18n.t("quote_checks.metadata").merge(
-        "gestes" => QuoteCheck::GESTE_TYPES_GROUPS.map do |group, values|
+        "gestes" => GesteTypes::GROUPS.map do |group, values|
           {
             group: group,
             values: values.map { I18n.t("quote_checks.geste_type.title.#{it}") }
