+++ lib/data_sources/data_ademe.rb
@@ -2,19 +2,23 @@
 
 require "net/http"
 
-# API Documentation: https://data.ademe.fr/data-fair/api/v1/datasets/liste-des-entreprises-rge-2/api-docs.json
-# Previously called "recensement des professionnels RGE (Reconnu Garant de l'Environnement)"
 # Now on Data.gouv https://www.data.gouv.fr/fr/dataservices/api-professionnels-rge/
 class DataAdeme
   class ServiceUnavailableError < StandardError; end
+
+  API_HOST = "https://data.ademe.fr/data-fair/api/v1"
+
+  def self.rge_openapi_uri
+    "#{API_HOST}/datasets/liste-des-entreprises-rge-2/api-docs.json" # V1 does not have OpenAPI schema
+  end
 
   # params: hash
   #   - qs: query string
   #   like "siret=12345678900000"
   #   or siret:%12345678900000 AND date_debut:[* TO 2023-01-13] AND date_fin:[2023-01-13 TO *]&
   def historique_rge_uri(params)
-    # https://data.ademe.fr/datasets/historique-rge V1
-    "https://data.ademe.fr/data-fair/api/v1/datasets/historique-rge/lines?#{params.compact.to_query}"
+    # Using V1 https://data.ademe.fr/datasets/historique-rge
+    "#{API_HOST}/datasets/historique-rge/lines?#{params.compact.to_query}"
 
     # New version V2 is not working https://data.ademe.fr/datasets/liste-des-entreprises-rge-2
     # Example: https://data.ademe.fr/data-fair/api/v1/datasets/liste-des-entreprises-rge-2/lines?page=1&after=1&size=12&sort=nom_entreprise&select=siret,nom_entreprise,adresse,code_postal,commune,latitude,longitude,telephone,email,site_internet,code_qualification,nom_qualification,url_qualification,nom_certificat,domaine,meta_domaine,organisme,particulier,_file.content,_file.content_type,_file.content_length,_attachment_url,_geopoint,_id,_i,_rand&format=json&q=12345678900000&q_mode=simple
