+++ lib/tools/clam_av.rb
