+++ lib/tools/error_notifier.rb
