+++ lib/tools/fiability.rb
