+++ lib/tools/tracking_hash.rb
