+++ lib/tools/uri_extended.rb
