+++ spec/lib/rge_validator_spec.rb
@@ -3,39 +3,6 @@
 require "rails_helper"
 
 RSpec.describe RgeValidator, type: :service do
-  describe ".ademe_geste_types" do
-    subject(:ademe_geste_types) { described_class.ademe_geste_types(domaine:, nom_certificat:) }
-
-    context "with only domaine" do
-      let(:domaine) { "Ventilation mécanique" }
-      let(:nom_certificat) { nil }
-
-      it "returns the corresponding MDSO geste type" do
-        expect(ademe_geste_types).to eq(%w[vmc_double_flux vmc_simple_flux])
-      end
-    end
-
-    context "with only nom_certificat" do
-      let(:domaine) { nil }
-      let(:nom_certificat) { "Chauffage +" }
-
-      it "returns the corresponding MDSO geste type" do
-        expect(ademe_geste_types).to eq(%w[chauffe_eau_solaire_individuel chauffe_eau_thermo])
-      end
-    end
-
-    context "with both domaine and nom_certificat" do
-      let(:domaine) { "Ventilation mécanique" }
-      let(:nom_certificat) { "Chauffage +" }
-
-      it "returns the corresponding MDSO geste type" do
-        expect(ademe_geste_types).to eq(
-          %w[chauffe_eau_solaire_individuel chauffe_eau_thermo vmc_double_flux vmc_simple_flux]
-        )
-      end
-    end
-  end
-
   describe ".geste_types_for_rge" do
     subject(:geste_types_for_rge) { described_class.geste_types_for_rge(siret, rge) }
 
@@ -44,15 +11,6 @@ RSpec.describe RgeValidator, type: :service do
 
     it "returns the corresponding MDSO geste types" do
       expect(geste_types_for_rge).to include("pac_hybride", "vmc_double_flux")
-    end
-  end
-
-  describe ".geste_types_with_certification" do
-    it "returns all MDSO geste types with certification" do
-      expect(described_class.geste_types_with_certification).to include(
-        "isolation_thermique_par_exterieur_ITE",
-        "chauffe_eau_solaire_individuel"
-      )
     end
   end
