+++ spec/lib/tools/fiability_spec.rb
