+++ spec/lib/tools/tracking_hash_spec.rb
