+++ spec/lib/tools/uri_extended_spec.rb
